GetById product

+++ APIMarketList.Domain/Interface/Repositories/Base/IBaseRepository.cs
@@ -6,7 +6,7 @@ namespace APIMarketList.Domain.Interface.Repositories.Base
     {
         Task<TEntity> AddAsync(TEntity entity);
         Task<int> DeleteAsync(TEntity entity);
-        Task<TEntity?> Get(int id);
+        Task<TEntity?> Get(long id);
         Task<IList<TEntity>> Get();
         Task<int> UpdateAsync(TEntity entity);
     }
